requirements.txt
requests==2.31.0
pandas==2.2.2
streamlit==1.36.0
pyarrow==16.1.0
altair==5.1.2
kafka-python==2.0.2
